refactor: RDB 기반 RefreshToken 관련 코드 제거
- RDB 기반 RefreshToken 관련 엔티티 및 레포지토리 제거
- Redis 레포지토리 이름 변경
