+++ src/main/java/dev/wgrgwg/somniverse/member/service/AuthService.java
@@ -5,7 +5,7 @@ import dev.wgrgwg.somniverse.member.domain.Member;
 import dev.wgrgwg.somniverse.member.dto.request.LoginRequest;
 import dev.wgrgwg.somniverse.member.dto.response.TokenResponse;
 import dev.wgrgwg.somniverse.member.exception.MemberErrorCode;
-import dev.wgrgwg.somniverse.member.repository.RefreshTokenRedisRepository;
+import dev.wgrgwg.somniverse.member.repository.RefreshTokenRepository;
 import dev.wgrgwg.somniverse.security.jwt.provider.JwtProvider;
 import dev.wgrgwg.somniverse.security.userdetails.CustomUserDetails;
 import java.util.Optional;
@@ -22,7 +22,7 @@ public class AuthService {
 
     private final AuthenticationManager authenticationManager;
     private final JwtProvider jwtProvider;
-    private final RefreshTokenRedisRepository refreshTokenRedisRepository;
+    private final RefreshTokenRepository refreshTokenRepository;
     private final MemberService memberService;
 
     @Transactional
@@ -36,7 +36,7 @@ public class AuthService {
 
         TokenResponse tokenResponse = jwtProvider.generateToken(member);
 
-        refreshTokenRedisRepository.save(tokenResponse.refreshToken(), member.getId().toString());
+        refreshTokenRepository.save(tokenResponse.refreshToken(), member.getId().toString());
 
         return tokenResponse;
     }
@@ -47,28 +47,28 @@ public class AuthService {
             throw new CustomException(MemberErrorCode.INVALID_REFRESH_TOKEN);
         }
 
-        String memberId = refreshTokenRedisRepository.findMemberIdByToken(refreshToken)
+        String memberId = refreshTokenRepository.findMemberIdByToken(refreshToken)
             .orElseThrow(() -> new CustomException(MemberErrorCode.REFRESH_TOKEN_NOT_FOUND));
 
         Member member = memberService.findById(Long.parseLong(memberId));
 
         TokenResponse newTokens = jwtProvider.generateToken(member);
 
-        refreshTokenRedisRepository.delete(refreshToken);
-        refreshTokenRedisRepository.save(newTokens.refreshToken(), memberId);
+        refreshTokenRepository.delete(refreshToken);
+        refreshTokenRepository.save(newTokens.refreshToken(), memberId);
 
         return new TokenResponse(newTokens.accessToken(), newTokens.refreshToken());
     }
 
     @Transactional
     public void logout(String refreshToken) {
-        Optional<String> memberIdOpt = refreshTokenRedisRepository.findMemberIdByToken(
+        Optional<String> memberIdOpt = refreshTokenRepository.findMemberIdByToken(
             refreshToken);
 
         if (memberIdOpt.isEmpty()) {
             throw new CustomException(MemberErrorCode.REFRESH_TOKEN_NOT_FOUND);
         }
 
-        refreshTokenRedisRepository.delete(refreshToken);
+        refreshTokenRepository.delete(refreshToken);
     }
 }
